requirements.txt
flask==2.3.2
langchain==0.0.304
langchain-community==0.0.6
sentence-transformers==2.2.2
faiss-cpu==1.7.4
requests==2.31.0
numpy==1.24.3
